+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Unreleased
 
+- [added] Implemented the `UpdateUserAsync()` API.
 - [added] Implemented the `CreateUserAsync()` and `UserRecordArgs` APIs.
+
+# v1.6.0
+
+- [added] `WebpushFcmOptions` added to the `WebpushConfig` class.
 - [added] Implemented the `GetUserByEmailAsync()` and `GetUserByPhoneNumberAsync()`
   APIs in the `FirebaseAuth` class.